+++ src/lib/pages/proposal-details/components/VoteQuorumText.tsx
@@ -1,12 +1,13 @@
 import { Text } from "@chakra-ui/react";
 
+import type { Ratio } from "lib/types";
 import { ProposalStatus } from "lib/types";
 import { formatPrettyPercent } from "lib/utils";
 
 interface VoteQuorumTextProps {
   status: ProposalStatus;
-  quorum: number;
-  totalVotes: Big;
+  quorum: Ratio<number>;
+  totalVotes: Ratio<number>;
   isCompact: boolean;
 }
 
@@ -29,7 +30,7 @@ export const VoteQuorumText = ({
 }: VoteQuorumTextProps) => {
   const fontVariant = isCompact ? "body2" : "body1";
   const quorumPercent = formatPrettyPercent(quorum);
-  const isPassingQuorum = totalVotes.gte(quorum);
+  const isPassingQuorum = totalVotes >= quorum;
 
   if (status === ProposalStatus.VOTING_PERIOD)
     return isPassingQuorum ? (