+++ src/lib/pages/validators/components/validators-table/ValidatorsTableMobileCard.tsx
@@ -6,12 +6,14 @@ import { MobileCardTemplate, MobileLabel } from "lib/components/table";
 import { ValidatorBadge } from "lib/components/ValidatorBadge";
 import type {
   Option,
+  Ratio,
   Token,
   TokenWithValue,
   U,
   ValidatorData,
 } from "lib/types";
 import {
+  divWithDefault,
   formatPrettyPercent,
   formatUTokenWithPrecision,
   getTokenLabel,
@@ -39,7 +41,7 @@ export const ValidatorsTableMobileCard = ({
     <MobileCardTemplate
       onClick={() =>
         navigate({
-          pathname: "/blocks/[validatorAddress]",
+          pathname: "/validators/[validatorAddress]",
           query: { validatorAddress: validator.validatorAddress },
         })
       }
@@ -76,7 +78,11 @@ export const ValidatorsTableMobileCard = ({
               <MobileLabel label="Voting Power" />
               <Text variant="body2" color="text.main">
                 {formatPrettyPercent(
-                  validator.votingPower.div(totalVotingPower).toNumber(),
+                  divWithDefault(
+                    validator.votingPower,
+                    totalVotingPower,
+                    0
+                  ).toNumber() as Ratio<number>,
                   2,
                   true
                 )}
@@ -102,7 +108,11 @@ export const ValidatorsTableMobileCard = ({
                 color={isZeroUptime ? "error.main" : "text.main"}
                 fontWeight={isZeroUptime ? 700 : undefined}
               >
-                {formatPrettyPercent((validator.uptime ?? 0) / 100, 0, true)}
+                {formatPrettyPercent(
+                  ((validator.uptime ?? 0) / 100) as Ratio<number>,
+                  0,
+                  true
+                )}
               </Text>
             </Flex>
           </Flex>